Migration, Annoucement, checkin, and all design files

+++ app/Http/Controllers/Business/StaffController.php
@@ -50,10 +50,14 @@ class StaffController extends Controller
         if ($request->hasFile('files')){   
             $image = $request->file('files')->store('staff');
         }
-        echo
         $random_password = Str::random(8);
+        // do {
+        //     $unique_code = rand(1000, 9999);
+        // } while (BusinessStaff::where('unique_code', $unique_code)->exists());
+    
+        $unique_code = generateUniqueCode();//my code 
         $company = $request->current_company;
-        $create = BusinessStaff::create(['business_id' => $company->id,'first_name' => $request->first_name, 'last_name' => $request->last_name, 'gender' => $request->gender,'position' =>$request->position, 'phone' => $request->phone ,'email'=>$request->email, 'profile_pic' => $image, 'bio'=> $request->bio,'address'=>$request->address,'city'=> $request->city,'state'=> $request->state,'postcode'=>$request->postcode,'birthdate'=>date('Y-m-d',strtotime($request->birthdate)) ,'password'=>Hash::make($random_password),'buddy_key'=>$random_password]);
+        $create = BusinessStaff::create(['business_id' => $company->id,'first_name' => $request->first_name, 'last_name' => $request->last_name, 'gender' => $request->gender,'position' =>$request->position, 'phone' => $request->phone ,'email'=>$request->email, 'profile_pic' => $image, 'bio'=> $request->bio,'address'=>$request->address,'city'=> $request->city,'state'=> $request->state,'postcode'=>$request->postcode,'birthdate'=>date('Y-m-d',strtotime($request->birthdate)) ,'password'=>Hash::make($random_password),'buddy_key'=>$random_password,   'unique_code' => $unique_code]);
         if($request->has('fromservice')){
             if($create){
                 return "success";
@@ -104,6 +108,20 @@ class StaffController extends Controller
     public function update(Request $request ,$business_id, $id)
     {
         //print_r($request->all());
+        // $uniqueCodeExists = BusinessStaff::where('unique_code', $request->unique_code)
+        // ->where('id', '!=', $id)
+        // ->exists();
+
+        // if ($uniqueCodeExists) {
+        //     return redirect()->back()->withErrors(['unique_code' => 'The unique code already exists.']);
+        // }
+        $uniqueCodeExists = BusinessStaff::where('unique_code', $request->unique_code)
+        ->where('id', '!=', $id)
+        ->exists();
+        $uniqueCodeExistsInUsers = User::where('unique_code', $request->unique_code)->exists();
+        if ($uniqueCodeExists || $uniqueCodeExistsInUsers) {
+            return redirect()->back()->withErrors(['unique_code' => 'The unique code already exists.']);
+        }
         if($request->has('image')){
             $image = $request->file('image')->store('staff');
             Storage::delete($request->oldImage);
@@ -135,7 +153,8 @@ class StaffController extends Controller
             'password'=>$password,
             'buddy_key'=>$buddy_key,
             'postcode'=>$request->postcode,
-            'birthdate'=>date('Y-m-d',strtotime($request->birthdate))
+            'birthdate'=>date('Y-m-d',strtotime($request->birthdate)),
+            'unique_code'=>$request->unique_code,
         ];   
         $staff->update($update);
         //print_r($update);exit;
